+++ lib/api/parser.dart
@@ -1,14 +1,11 @@
 import 'package:light_html_editor/api/regex_provider.dart';
 import 'package:light_html_editor/api/richtext_node.dart';
-import 'package:light_html_editor/placeholder.dart';
+import 'package:light_html_editor/api/v2/node_v2.dart';
+import 'package:light_html_editor/light_html_editor.dart';
 
-///
-/// used to transform written HTML-Richtext into a parse-tree and holds a number
-/// of utility-methods like stripping tags from a text to access the tagless text
-///
 class Parser {
   /// remove all tags from a HTML-Richtext and return the raw, unformatted text
-  String cleanTagsFromRichtext(String text) {
+  static String cleanTagsFromRichtext(String text) {
     RegExpMatch? match = RegExProvider.tagRegex.firstMatch(text);
 
     while (match != null) {
@@ -23,7 +20,7 @@ class Parser {
   /// substitutes the variables occuring in [text] with their respective values
   /// stored in [placeholders]
   ///
-  String replaceVariables(
+  static String replaceVariables(
     String text, {
     List<RichTextPlaceholder> placeholders = const [],
     String placeholderMarker = "\\\$",
@@ -41,52 +38,57 @@ class Parser {
     return text;
   }
 
-  /// Parse HTML-Richtext into a ParseTree that holds all information needed
-  /// to create a Flutter widget out of it
-  DocumentNode parse(String text) {
-    DocumentNode currentNode = DocumentNode(
-      Tag("", {}, true, 0),
-      null,
-    );
+  NodeV2 parse(String? rawText) {
+    NodeV2 root = NodeV2.root();
+    if (rawText == null) return root;
 
-    String remainingText = text;
+    _parseAndAddTo(root, rawText);
 
+    return root;
+  }
+
+  String _parseAndAddTo(NodeV2 k, String remainingText) {
     while (remainingText.isNotEmpty) {
       RegExpMatch? nextTagMatch =
           RegExProvider.tagRegex.firstMatch(remainingText);
 
+      // no tags left, just plaintext
       if (nextTagMatch == null) {
-        currentNode.text.add(remainingText);
+        k.children.add(SimpleNode(k, remainingText));
         remainingText = "";
         break;
       }
 
-      Tag tag = Tag.decodeTag(
-          remainingText.substring(nextTagMatch.start, nextTagMatch.end));
-
-      String tagName, leading;
-
-      if (tag.isStart) {
-        tagName = tag.name;
-
-        leading = remainingText.substring(0, nextTagMatch.start);
-        remainingText = remainingText.substring(nextTagMatch.end);
-
-        currentNode.text.add(leading);
-        currentNode = DocumentNode(tag, currentNode);
+      // a tag has been found - was it at string-start though?
+      if (nextTagMatch.start > 0) {
+        k.children.add(
+          SimpleNode(k, remainingText.substring(0, nextTagMatch.start)),
+        );
+        remainingText = remainingText.substring(nextTagMatch.start);
       } else {
-        tagName = tag.name;
+        Tag tag = Tag.decodeTag(
+            remainingText.substring(nextTagMatch.start, nextTagMatch.end));
 
-        if (tagName != currentNode.scope.name) return currentNode;
+        if (tag.isStart) {
+          // encountered a start-tag
+          NodeV2 nodeNew = NodeV2.fromTag(k, tag);
+          k.children.add(nodeNew);
 
-        leading = remainingText.substring(0, nextTagMatch.start);
-        remainingText = remainingText.substring(nextTagMatch.end);
+          remainingText = remainingText.substring(tag.rawTagLength);
+          remainingText = _parseAndAddTo(nodeNew, remainingText);
+        } else {
+          // encountered an end-tag
+          if (k.tag != tag.name) {
+            throw Exception(
+                "Unexpected end-tag! Expected '</${k.tag}>' but found '</${tag.name}>'");
+          }
 
-        currentNode.text.add(leading);
-        if (currentNode.parent != null) currentNode = currentNode.parent!;
+          remainingText = remainingText.substring(tag.rawTagLength);
+          return remainingText;
+        }
       }
     }
 
-    return currentNode.root;
+    return remainingText;
   }
 }