+++ lib/api/text_renderer.dart
@@ -2,18 +2,18 @@ import 'dart:math';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:light_html_editor/api/richtext_node.dart';
+import 'package:light_html_editor/api/v2/node_v2.dart';
 import 'package:light_html_editor/data/text_constants.dart';
 import 'package:light_html_editor/light_html_editor.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 ///
-/// Used to conver the tree of [DocumentNode] nodes into a list of [RichText]
+/// Used to conver the tree of [NodeV2] nodes into a list of [RichText]
 /// which resemble the paragraphs of the text
 ///
 class TextRenderer {
   /// root of the tree to be converted
-  final DocumentNode root;
+  final String rawHtml;
 
   /// renderer settings to be applied
   final RendererDecoration rendererDecoration;
@@ -55,6 +55,7 @@ class TextRenderer {
     }
 
     return RichText(
+      key: UniqueKey(),
       maxLines: maxLines,
       text: TextSpan(
         children: content,
@@ -67,7 +68,7 @@ class TextRenderer {
 
   /// recursively parses the richtext-tree
   TextRenderer(
-    this.root,
+    this.rawHtml,
     this.rendererDecoration,
     this.maxLength,
     this.maxLines,
@@ -75,30 +76,9 @@ class TextRenderer {
     this.placeholderMarker,
     this.ignoreLinebreaks,
   ) {
-    for (RendererTextProperties textProperties
-        in rendererDecoration.textProperties) {
-      _fontSizes[textProperties.tagName] = textProperties.fontSize;
-      if (textProperties.fontFamily != null)
-        _fontFamilies[textProperties.tagName] = textProperties.fontFamily!;
-    }
+    _initFontStyles();
 
-    if (!_fontSizes.containsKey("")) {
-      _fontSizes[""] = TextConstants.defaultFontSize;
-    }
-
-    if (!_fontFamilies.containsKey("")) {
-      _fontFamilies[""] = TextConstants.defaultFontFamily;
-    }
-
-    for (String tag in TextConstants.headerSizes.keys) {
-      if (!_fontSizes.containsKey(tag))
-        _fontSizes[tag] = TextConstants.headerSizes[tag]!;
-
-      if (!_fontFamilies.containsKey(tag)) {
-        _fontFamilies[tag] = TextConstants.defaultFontFamily;
-      }
-    }
-
+    NodeV2 root = Parser().parse(rawHtml);
     _proccessNode(root);
 
     int textLength = 0;
@@ -115,7 +95,7 @@ class TextRenderer {
         full = true;
       }
 
-      nodeText = Parser().replaceVariables(
+      nodeText = Parser.replaceVariables(
         nodeText,
         placeholders: placeholders,
         placeholderMarker: placeholderMarker,
@@ -157,6 +137,35 @@ class TextRenderer {
     _performLinebreak();
   }
 
+  ///
+  /// fills in the default values for font-size and font-family
+  ///
+  void _initFontStyles() {
+    for (RendererTextProperties textProperties
+        in rendererDecoration.textProperties) {
+      _fontSizes[textProperties.tagName] = textProperties.fontSize;
+      if (textProperties.fontFamily != null)
+        _fontFamilies[textProperties.tagName] = textProperties.fontFamily!;
+    }
+
+    if (!_fontSizes.containsKey("")) {
+      _fontSizes[""] = TextConstants.defaultFontSize;
+    }
+
+    if (!_fontFamilies.containsKey("")) {
+      _fontFamilies[""] = TextConstants.defaultFontFamily;
+    }
+
+    for (String tag in TextConstants.headerSizes.keys) {
+      if (!_fontSizes.containsKey(tag))
+        _fontSizes[tag] = TextConstants.headerSizes[tag]!;
+
+      if (!_fontFamilies.containsKey(tag)) {
+        _fontFamilies[tag] = TextConstants.defaultFontFamily;
+      }
+    }
+  }
+
   void _performLinebreak() {
     if (_currentParagraph.length > 0) _paragraphs.add(_currentParagraph);
     _currentParagraph = [];
@@ -166,7 +175,7 @@ class TextRenderer {
   /// returns the underline-value for this node. checks first if the node is
   /// descendant of a link-node and looks for <u> tags afterwards
   ///
-  TextDecoration _underline(DocumentNode node) {
+  TextDecoration _textDecoration(SimpleNode node) {
     if (node.isLink) {
       if (rendererDecoration.linkUnderline != null)
         return rendererDecoration.linkUnderline!
@@ -176,14 +185,14 @@ class TextRenderer {
         return TextConstants.defaultLinkUnderline;
     }
 
-    return node.underline;
+    return node.textDecoration;
   }
 
   ///
   /// returns the underline-value for this node. checks first if the node is
   /// descendant of a link-node and looks for <u> tags afterwards
   ///
-  Color _color(DocumentNode node) {
+  Color _color(SimpleNode node) {
     if (node.isLink) {
       if (rendererDecoration.linkColor != null)
         return rendererDecoration.linkColor!;
@@ -191,38 +200,38 @@ class TextRenderer {
         return TextConstants.defaultLinkColor;
     }
 
-    return node.textColor != null
-        ? node.textColor!
-        : rendererDecoration.defaultColor;
+    return node.textColor ?? rendererDecoration.defaultColor;
   }
 
   ///
   /// converts a subtree into a list of [_TextNode], resulting into an in-order
   /// flattening of the subtree.
   ///
-  void _proccessNode(DocumentNode node) {
-    for (int i = 0; i < node.text.length; i++) {
-      if (node.text[i].isNotEmpty) {
-        _flattenedNodes.add(
-          _TextNode(
-            node.text[i],
-            TextStyle(
-              fontSize: node.fontSize(_fontSizes) != null
-                  ? node.fontSize(_fontSizes)
-                  : _fontSizes[""],
-              fontFamily: _fontFamilies[""],
-              fontWeight: node.isBold ? FontWeight.bold : FontWeight.normal,
-              fontStyle: node.isItalics ? FontStyle.italic : FontStyle.normal,
-              color: _color(node),
-              decoration: _underline(node),
-            ),
-            node.invokesNewline && (i == 0),
-            node.invokesNewline && (i == node.text.length - 1),
-            node.linkTarget,
+  void _proccessNode(NodeV2 node) {
+    if (node is SimpleNode) {
+      bool addLineBreakBefore = node.isHeader || node.isParagraph;
+      bool addLineBreakAfter = node.isHeader || node.isParagraph;
+
+      _flattenedNodes.add(
+        _TextNode(
+          node.body,
+          TextStyle(
+            fontSize: node.fontSize(_fontSizes),
+            fontFamily: node.fontFamily ?? _fontFamilies[""],
+            fontWeight: node.fontWeight,
+            fontStyle: FontStyle.normal,
+            color: _color(node),
+            decoration: _textDecoration(node),
           ),
-        );
-      }
-      if (i < node.children.length) _proccessNode(node.children[i]);
+          addLineBreakBefore,
+          addLineBreakAfter,
+          node.linkTarget,
+        ),
+      );
+    }
+
+    for (NodeV2 child in node.children) {
+      _proccessNode(child);
     }
   }
 }