+++ lib/light_html_editor.dart
@@ -1,7 +1,7 @@
 export 'editor.dart';
 export 'placeholder.dart';
 export 'renderer.dart';
-export 'api/parser.dart';
 export 'data/editor_properties.dart';
 export 'data/renderer_decoration.dart';
 export 'data/renderer_text_properties.dart';
+export 'api/parser.dart';